Move AsmConstWalker from wasm-linker to wasm-emscripten
